refactor: replace useAuthStore usage in stores
