+++ src/lib/hooks/useAuth.ts
@@ -1,0 +1,1 @@
+export * from '../../hooks/auth/useAuth';
